Fixed a bug in the NavigatorAgent
The NavigatorAgent was getting the map before the map was instantiated as a GameObject.
To fix this, a check was added that checks if the map is null. If it is, it tries to get the map again. It if fails twice, it shortcircuits the function to prevent an error.