Making GameObject children of the map
GameObject.cs:
the children GameObjects are now updated/onresized/rendered through the parent GameObject.
OnDestroy, the GameObject's reference is removed from the parent.
Instantiate can take in a reference to a GameObject as a parameter and will set it to the parent.

Transform.cs:
A new function, DestroyParentReference() was created to remove the Tranform reference from the parent.

GameManager.cs:
The player's object is now a child of the GameManager object.

All objects created in the map now are instantiated with the map as there parent GameObject.